Move utility bits to the utility class
