+++ SpiNNaker-comms/src/main/java/uk/ac/manchester/spinnaker/alloc/client/ClientSession.java
@@ -25,12 +25,12 @@ import static java.util.Objects.nonNull;
 import static java.util.Objects.requireNonNull;
 import static org.apache.commons.io.IOUtils.readLines;
 import static org.slf4j.LoggerFactory.getLogger;
+import static uk.ac.manchester.spinnaker.alloc.client.ClientUtils.asDir;
 import static uk.ac.manchester.spinnaker.alloc.client.ProxyProtocol.CLOSE;
 import static uk.ac.manchester.spinnaker.alloc.client.ProxyProtocol.MSG;
 import static uk.ac.manchester.spinnaker.alloc.client.ProxyProtocol.MSG_TO;
 import static uk.ac.manchester.spinnaker.alloc.client.ProxyProtocol.OPEN;
 import static uk.ac.manchester.spinnaker.alloc.client.ProxyProtocol.OPEN_U;
-import static uk.ac.manchester.spinnaker.alloc.client.SpallocClientFactory.asDir;
 import static uk.ac.manchester.spinnaker.alloc.client.SpallocClientFactory.checkForError;
 import static uk.ac.manchester.spinnaker.alloc.client.SpallocClientFactory.readJson;
 import static uk.ac.manchester.spinnaker.alloc.client.SpallocClientFactory.writeForm;
