+++ SpiNNaker-comms/src/main/java/uk/ac/manchester/spinnaker/alloc/client/ClientUtils.java
@@ -20,7 +20,10 @@ import static java.util.Objects.isNull;
 import static java.util.concurrent.TimeUnit.MILLISECONDS;
 
 import java.net.SocketTimeoutException;
+import java.net.URI;
 import java.nio.ByteBuffer;
+import java.util.List;
+import java.util.Objects;
 import java.util.concurrent.BlockingQueue;
 
 /** Shared helper because we can't use a superclass. */
@@ -48,5 +51,37 @@ abstract class ClientUtils {
 			throw new SocketTimeoutException();
 		}
 		return msg;
+	}
+
+	/**
+	 * Add a {@code /} to the end of the path part of a URI.
+	 *
+	 * @param uri
+	 *            The URI to amend. Assumed to be HTTP or HTTPS.
+	 * @return The amended URI.
+	 */
+	static URI asDir(URI uri) {
+		var path = uri.getPath();
+		if (!path.endsWith("/")) {
+			path += "/";
+			uri = uri.resolve(path);
+		}
+		return uri;
+	}
+
+	/**
+	 * Make a read-only shallow copy of a list.
+	 *
+	 * @param <T>
+	 *            The type of elements in the list. These are
+	 *            <em>recommended</em> to be immutable, but this is not
+	 *            enforced.
+	 * @param list
+	 *            The list to make a read-only copy of. The elements in the list
+	 *            are not themselves copied.
+	 * @return The read-only copy.
+	 */
+	static <T> List<T> readOnlyCopy(List<T> list) {
+		return Objects.isNull(list) ? List.of() : List.copyOf(list);
 	}
 }
